requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
billiard==3.3.0.23
celery==3.1.23
decorator==4.0.10
Django==1.8.2
django-anymail==0.4.2
django-extensions==1.6.7
djangorestframework==3.4.0
httplib2==0.9.2
ipython==5.0.0
ipython-genutils==0.1.0
kombu==3.0.35
pexpect==4.2.0
pickleshare==0.7.3
prompt-toolkit==1.0.3
psycopg2==2.6.1
ptyprocess==0.5.1
Pygments==2.1.3
python-dateutil==2.5.3
pytz==2016.6.1
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
